휴게소 관련 모든 코드 리팩토링 완료
