+++ backend_core/src/main/java/com/mini/mini_2/facility/domain/entity/FacilityEntity.java
@@ -20,7 +20,6 @@ import lombok.ToString;
 
 @Entity
 @Table(name = "facility")
-
 @Builder
 @Setter
 @Getter
