+++ backend_core/src/main/java/com/mini/mini_2/favorite/application/dto/FavoriteRequestDTO.java
@@ -1,0 +1,33 @@
+package com.mini.mini_2.favorite.application.dto; // [수정] 패키지 변경
+
+import com.mini.mini_2.favorite.domain.entity.FavoriteEntity;
+import com.mini.mini_2.rest_area.domain.entity.RestAreaEntity;
+import com.mini.mini_2.user.domain.entity.UserEntity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+@Builder
+@Setter
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class FavoriteRequestDTO {
+    private Integer userId;
+    private Integer restAreaId;
+    
+    private String description;
+    
+    public FavoriteEntity toEntity(UserEntity userEntity, RestAreaEntity restAreaEntity) {
+        return FavoriteEntity.builder()
+                             .user(userEntity)
+                             .restArea(restAreaEntity)
+                             .description(this.description)
+                             .build();
+    }
+}
