+++ backend_core/src/main/java/com/mini/mini_2/favorite/domain/entity/FavoriteEntity.java
@@ -26,7 +26,7 @@ import lombok.ToString;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-@ToString(exclude = {"restArea"})
+@ToString(exclude = {"restArea", "user"}) // [수정] user도 exclude
 public class FavoriteEntity {
     
     @Id
