--- backend_core/src/main/java/com/mini/mini_2/review/repository/ReviewRepository.java
@@ -1,23 +1,0 @@
-package com.mini.mini_2.review.repository;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.mini.mini_2.review.domain.entity.ReviewEntity;
-
-
-
-@Repository
-public interface ReviewRepository extends JpaRepository<ReviewEntity, Integer> {
-    
-    // 기본 조회
-    List<ReviewEntity> findByRestArea_RestAreaId(Integer restAreaId);
-    List<ReviewEntity> findByUser_UserId(Integer userId);
-
-    // 최신순(정렬)
-    List<ReviewEntity> findByRestArea_RestAreaIdOrderByCreatedAtDesc(Integer restAreaId);
-    // 평점순(정렬)
-    List<ReviewEntity> findByRestArea_RestAreaIdOrderByRatingDesc(Integer restAreaId);
-}
