+++ backend_core/src/test/java/com/mini/mini_2/all/TestAllTables.java
@@ -6,14 +6,14 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.Rollback;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.mini.mini_2.facility.domain.dto.FacilityRequestDTO;
-import com.mini.mini_2.facility.domain.dto.FacilityResponseDTO;
+import com.mini.mini_2.facility.application.dto.FacilityRequestDTO;
+import com.mini.mini_2.facility.application.dto.FacilityResponseDTO;
 import com.mini.mini_2.facility.domain.entity.FacilityEntity;
-import com.mini.mini_2.facility.repository.FacilityRepository;
-import com.mini.mini_2.favorite.domain.dto.FavoriteRequestDTO;
-import com.mini.mini_2.favorite.domain.dto.FavoriteResponseDTO;
+import com.mini.mini_2.facility.domain.FacilityRepository;
+import com.mini.mini_2.favorite.application.dto.FavoriteRequestDTO;
+import com.mini.mini_2.favorite.application.dto.FavoriteResponseDTO;
 import com.mini.mini_2.favorite.domain.entity.FavoriteEntity;
-import com.mini.mini_2.favorite.repository.FavoriteRepository;
+import com.mini.mini_2.favorite.domain.FavoriteRepository;
 import com.mini.mini_2.food.application.dto.FoodRequestDTO;
 import com.mini.mini_2.food.application.dto.FoodResponseDTO;
 import com.mini.mini_2.food.domain.FoodRepository;
